Showing namespaces on each namespace page.
This commit builds on top of PR https://github.com/fabpot/Sami/pull/13,
but it has been updated for recent changes to Sami. In addition to
some minor tweaks, this PR shows child namespaces as a comma separated
list of namespaces rather than in a table to take up less space. This
allows users to see classes and interfaces on a page without having
to scroll down past the fold (providing a better user experience IMO).

Closes #13. Closes #119.