Using typeahead and bloodhound for search.
This commit updates the search engine to use Twitter's
typeahead and Bloodhound plugins to implement the search
engine and provide typeahead suggestions when entering
search terms. Selecting an auto-completed term will take you
directly to the page, while hitting enter on the form will
take you to the search page. Both the search page and
auto-complete searches use bloodhound for the search engine.

A "Search" link was added to the menu so that mobile users
will have a search function when the left-nav disappears.